requirements.txt
requests~=2.28.2
beautifulsoup4
Pillow~=9.5.0